Fix MCP mode to match CLI/SDK behavior by eliminating code duplication
Previously, MCP mode had different behavior than CLI/SDK mode because it:
- Failed to remove <thinking> tags from responses
- Lacked attempt_complete recovery logic
- Used different XML parsing pipeline

This fix makes MCP mode a true thin wrapper by:

## Changes Made

1. **Created shared utilities** (`xmlParsingUtils.js`):
   - Extracted thinking tag removal logic
   - Extracted attempt_complete recovery patterns
   - Made utilities reusable by both modes

2. **Eliminated code duplication**:
   - Removed ~80 lines of duplicated code from MCP xmlBridge.js
   - Refactored CLI/SDK tools.js to use shared utilities
   - Both modes now use identical core logic

3. **Fixed circular dependencies**:
   - Created xmlParsingUtils.js as shared dependency
   - Clean import structure with no circular references

## Impact

- MCP mode now behaves identically to CLI/SDK mode
- <thinking> tags properly removed in both modes
- attempt_complete recovery works consistently
- Single source of truth for XML parsing logic
- Improved maintainability and consistency
